ver: 1.5 1、清理内存泄漏。 2、修复CountMutex使用错误可能带来的死锁BUG。 3、增加IAccessible接口，为linux, macos系统实现该接口。 4、基于fun_test的单元测试，修复测试发现的API实现BUG。 5、让GDI支持字体回退。 6、以 fcntl POSIX 记录锁实现，取代 XSI System V 信号量，解决进程异常退出后，程序可能无法启动的问题。

+++ src/gdi/cairo/builtin_image.cpp
@@ -28,7 +28,7 @@ struct MemBlock
 static cairo_status_t read_png_from_memory(void *closure, unsigned char *data, unsigned int length)
 {
     MemBlock *memBlock = (MemBlock *)closure;
-    if (length > memBlock->bufLen - memBlock->pos)
+    if (length > (unsigned int)(memBlock->bufLen - memBlock->pos))
     {
         return CAIRO_STATUS_READ_ERROR;
     }
